requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
google-generativeai==0.3.0
requests==2.31.0
python-decouple==3.8
feedparser==6.0.10
beautifulsoup4==4.12.2
django-extensions==3.2.3
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn